cleanup

+++ mapel/voting/objects/Election.py
@@ -48,7 +48,7 @@ class Election(Instance):
 
     def votes_to_potes(self) -> np.ndarray:
         """ Convert votes to positional votes """
-        potes = np.zeros([self.num_voters, self.num_candidates])
+        potes = np.zeros([self.num_voters, self.num_candidates], dtype=int)
         for i in range(self.num_voters):
             for j in range(self.num_candidates):
                 potes[i][self.votes[i][j]] = j
